chore: refactor to track userStore as dep
