+++ src/pages/Research/Content/ResearchUpdate.tsx
@@ -14,7 +14,7 @@ import { ResearchComments } from './ResearchComments/ResearchComments'
 import styled from '@emotion/styled'
 import type { IComment } from 'src/models'
 import { Link, useNavigate } from 'react-router-dom'
-import { useContributorsData } from 'src/common/hooks/contributorsData'
+import { useContributorsData } from 'src/common/hooks/getContributorsData'
 import { useResearchStore } from 'src/stores/Research/research.store'
 import { useCommonStores } from 'src/index'
 import { formatImagesForGallery } from 'src/utils/formatImageListForGallery'
